refactor: rename input parameter to data in transformers and reorganize into packages
