+++ src/main/java/org/phong/zenflow/plugin/subdomain/executors/builtin/data/data_transformer/impl/string/ConcatTransformer.java
@@ -1,4 +1,4 @@
-package org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.impl;
+package org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.impl.string;
 
 import org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.exception.DataTransformerExecutorException;
 import org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.interfaces.DataTransformer;
@@ -14,11 +14,11 @@ public class ConcatTransformer implements DataTransformer {
     }
 
     @Override
-    public Object transform(Object input, Map<String, Object> params) {
+    public Object transform(Object data, Map<String, Object> params) {
         if (params == null || !params.containsKey("suffix")) {
             throw new DataTransformerExecutorException("Suffix parameter is required for concat transformer.");
         }
         String suffix = String.valueOf(params.get("suffix"));
-        return input + suffix;
+        return data + suffix;
     }
 }
