+++ src/main/java/org/phong/zenflow/plugin/subdomain/executors/builtin/data/data_transformer/impl/string/LowercaseTransformer.java
@@ -1,4 +1,4 @@
-package org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.impl;
+package org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.impl.string;
 
 import org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.interfaces.DataTransformer;
 import org.springframework.stereotype.Component;
@@ -13,10 +13,10 @@ public class LowercaseTransformer implements DataTransformer {
     }
 
     @Override
-    public Object transform(Object input, Map<String, Object> params) {
-        if (input == null) {
+    public Object transform(Object data, Map<String, Object> params) {
+        if (data == null) {
             return null;
         }
-        return input.toString().toLowerCase();
+        return data.toString().toLowerCase();
     }
 }
