+++ src/main/java/org/phong/zenflow/plugin/subdomain/executors/builtin/data/data_transformer/impl/string/SubstringTransformer.java
@@ -1,4 +1,4 @@
-package org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.impl;
+package org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.impl.string;
 
 import org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.interfaces.DataTransformer;
 import org.springframework.stereotype.Component;
@@ -11,11 +11,11 @@ public class SubstringTransformer implements DataTransformer {
     public String getName() { return "substring"; }
 
     @Override
-    public Object transform(Object input, Map<String, Object> params) {
-        if (input == null) {
+    public Object transform(Object data, Map<String, Object> params) {
+        if (data == null) {
             return null;
         }
-        String strInput = input.toString();
+        String strInput = data.toString();
         int start = (int) params.getOrDefault("start", 0);
         int end = (int) params.getOrDefault("end", strInput.length());
         return strInput.substring(start, end);
