+++ src/main/java/org/phong/zenflow/plugin/subdomain/executors/builtin/data/data_transformer/impl/text/RegexSplitTransformer.java
@@ -1,4 +1,4 @@
-package org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.impl;
+package org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.impl.text;
 
 import org.phong.zenflow.plugin.subdomain.executors.builtin.data.data_transformer.interfaces.DataTransformer;
 import org.springframework.stereotype.Component;
@@ -14,14 +14,14 @@ public class RegexSplitTransformer implements DataTransformer {
     }
 
     @Override
-    public Object transform(Object input, Map<String, Object> params) {
-        if (input == null) {
+    public Object transform(Object data, Map<String, Object> params) {
+        if (data == null) {
             return null;
         }
         String regex = (String) params.get("pattern");
         if (regex == null) {
             throw new IllegalArgumentException("Pattern parameter is required for regex_split transformer.");
         }
-        return Arrays.asList(input.toString().split(regex));
+        return Arrays.asList(data.toString().split(regex));
     }
 }
