+++ src/main/resources/db/migration/V8__insert_default_plugin_core_nodes.sql
@@ -295,7 +295,7 @@ VALUES
          "name": {
            "type": "string"
          },
-         "input": {},
+         "data": {},
          "params": {
            "type": "object",
            "default": {}
@@ -317,7 +317,7 @@ VALUES
          }
        },
        "required": [
-         "input"
+         "data"
        ],
        "if": {
          "properties": {
